Rest controllers

+++ src/main/java/com/miwi/carrental/controller/ManageController.java
@@ -1,34 +1,32 @@
 package com.miwi.carrental.controller;
 
 import com.miwi.carrental.domain.dto.CarDto;
-import com.miwi.carrental.domain.entity.BodyType;
-import com.miwi.carrental.domain.entity.CarModel;
-import com.miwi.carrental.domain.entity.CarStatus;
-import com.miwi.carrental.domain.entity.Location;
-import com.miwi.carrental.domain.view.CarViewAdmin;
+import com.miwi.carrental.domain.entity.Car;
 import com.miwi.carrental.service.entityservice.BodyTypeService;
 import com.miwi.carrental.service.entityservice.CarModelService;
 import com.miwi.carrental.service.entityservice.CarService;
 import com.miwi.carrental.service.entityservice.CarStatusService;
 import com.miwi.carrental.service.entityservice.LocationService;
 import com.miwi.carrental.service.viewservice.CarViewAdminService;
-import java.util.List;
-import java.util.Optional;
 import javax.validation.Valid;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-import org.springframework.stereotype.Controller;
-import org.springframework.ui.Model;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.MediaType;
+import org.springframework.http.ResponseEntity;
 import org.springframework.validation.BindingResult;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.ModelAttribute;
+import org.springframework.web.bind.annotation.CrossOrigin;
+import org.springframework.web.bind.annotation.DeleteMapping;
+import org.springframework.web.bind.annotation.PatchMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RequestMethod;
+import org.springframework.web.bind.annotation.RestController;
 
-@Controller
-@RequestMapping("/manage")
+@RestController
+@RequestMapping(path = "/manage", produces = MediaType.APPLICATION_JSON_VALUE)
+@CrossOrigin(origins = "*")
 public class ManageController {
 
   private Logger logger = LoggerFactory.getLogger(getClass().getName());
@@ -55,7 +53,7 @@ public class ManageController {
     this.carViewAdminService = carViewAdminService;
   }
 
-  @ModelAttribute
+  /*@ModelAttribute
   public void getLists(final Model model) {
     List<BodyType> bodyTypes = bodyTypeService.findAll();
     List<CarModel> carModels = carModelService.findAll();
@@ -66,17 +64,18 @@ public class ManageController {
     model.addAttribute("carModels", carModels);
     model.addAttribute("locations", locations);
     model.addAttribute("carStatuses", carStatuses);
-  }
+  }*/
 
-  @GetMapping("/new-car")
+  /*@GetMapping("/new-car")
   public String newCarForm(Model model) {
     CarDto newCar = new CarDto();
     model.addAttribute("car", newCar);
     return "carForm";
-  }
+  }*/
 
-  @PostMapping("/add-car")
-  public String createCar(@Valid @ModelAttribute("car") CarDto car, BindingResult bindingResult) {
+  @PostMapping(path = "/add-car", consumes = MediaType.APPLICATION_JSON_VALUE)
+  public ResponseEntity<Car> createCar(@Valid @RequestBody CarDto car,
+      BindingResult bindingResult) {
     if (bindingResult.hasErrors()) {
 
       System.out.println("there were errors");
@@ -86,28 +85,16 @@ public class ManageController {
             errorObjectName.substring(0, 1).toUpperCase() + errorObjectName.substring(1),
             error.getDefaultMessage());
       });
-      return "carForm";
+      return new ResponseEntity<>(null, HttpStatus.BAD_REQUEST);
     }
-    carService.createNewCar(car);
-    return "index";
+    Car newCar = carService.createNewCar(car);
+
+    return new ResponseEntity<>(newCar, HttpStatus.CREATED);
   }
 
-  @GetMapping("/edit-form/{id}")
-  public String editCarForm(Model model, @PathVariable("id") Optional<Long> idParam) {
-    if (idParam.isPresent()) {
-      Long idl = idParam.get();
-      if (carViewAdminService.findById(idl).isPresent()) {
-        CarViewAdmin car = carViewAdminService.findById(idl).get();
-        model.addAttribute("editCar", car);
-        model.addAttribute("carDto", new CarDto());
-        return "editCarForm";
-      }
-    }
-    return "redirect:/manage";
-  }
-
-  @PostMapping("/edit-car")
-  public String editCar(@Valid @ModelAttribute CarDto carDto, BindingResult bindingResult) {
+  @PatchMapping(path = "/edit-car/{id}", consumes = MediaType.APPLICATION_JSON_VALUE)
+  public ResponseEntity<Car> editCar(@Valid @RequestBody CarDto carDto,
+      BindingResult bindingResult) {
     if (bindingResult.hasErrors()) {
       System.out.println("there were errors");
       bindingResult.getAllErrors().forEach(error -> {
@@ -116,23 +103,29 @@ public class ManageController {
             errorObjectName.substring(0, 1).toUpperCase() + errorObjectName.substring(1),
             error.getDefaultMessage());
       });
-      return "editCarForm";
+      return new ResponseEntity<>(null, HttpStatus.BAD_REQUEST);
     }
-    carService.createNewCar(carDto);
-    return "index";
+    Car car = carService.createNewCar(carDto);
+    return new ResponseEntity<>(car, HttpStatus.OK);
   }
 
-  @RequestMapping(path = "/delete/{id}", method = {RequestMethod.DELETE, RequestMethod.GET})
-  public String deleteCarById(@PathVariable("id") Long id) {
+  @DeleteMapping(path = "/delete/{id}")
+  public ResponseEntity<Car> deleteCarById(@PathVariable("id") Long id) {
+    if (!carService.findById(id).isPresent()) {
+      return new ResponseEntity<>(null, HttpStatus.NOT_FOUND);
+    }
     carService.deleteById(id);
-    return "redirect:/cars/our-fleet";
+    return new ResponseEntity<>(null, HttpStatus.NO_CONTENT);
   }
 
-  @RequestMapping(path = "/make-availability/{carId}/{status}", method = {RequestMethod.POST,
-      RequestMethod.GET})
-  public String makeAvailability(@PathVariable("carId") Long carId,
+  @PatchMapping(path = "/make-availability/{carId}/{status}", consumes = MediaType.APPLICATION_JSON_VALUE)
+  public ResponseEntity<Car> makeAvailability(@PathVariable("carId") Long carId,
       @PathVariable("status") String carStatus) {
+    if (!carService.findById(carId).isPresent() || carStatus.isEmpty()) {
+      return new ResponseEntity<>(null, HttpStatus.NOT_FOUND);
+    }
+    Car car = carService.findById(carId).get();
     carService.changeToAvailable(carId, carStatus);
-    return "redirect:/cars/our-fleet";
+    return new ResponseEntity<>(car, HttpStatus.OK);
   }
 }
